am7xxx: switch from device quirks to device ops
Define device specific operations in order to allow supporting different
implementation of a given operation by different device models.

For instance this will be used to add an alternative implementation of
the 'set_power_mode' operation for the Philips/SagemCom PicoPix 2055
projector.

For now device specific callbacks are used only for the 'set_power_mode'
and 'set_zoom_mode' operations.